correções em algumas views e salvamento de fotos no bd

+++ routes/web.php
@@ -25,14 +25,13 @@ Route::get('/dashboard', [UserController::class, 'dashboard'])->middleware(['aut
 Route::delete('/delete/{id}', [UserController::class, 'delete'])->middleware(['auth', 'is_admin'])->name('user.delete');
 
 
-Route::get('/product/edit/{id}',[ProductController::class, 'edit'])->name('product.edit');
+Route::get('/product/edit/{id}',[ProductController::class, 'edit'])->middleware(['auth', 'is_admin'])->name('product.edit');
 Route::put('/product/edit/{id}',[ProductController::class, 'update'])->name('product.update');
 Route::get('/products', [ProductController::class, 'index'])->middleware(['auth', 'is_admin'])->name('product.index');
-
 Route::get('/product/create', [ProductController::class, 'create'])->name('product.create');
 Route::post('/product', [ProductController::class, 'store'])->name('product.store');
 Route::get('/product', [ProductController::class, 'search'])->name('search');
-Route::delete('/products/{id}', [ProductController::class, 'destroy'])->name('product.delete');
+Route::delete('/products/{id}', [ProductController::class, 'destroy'])->middleware(['auth', 'is_admin'])->name('product.delete');
 Route::get('/products/{id}',[ProductController::class, 'show'])->middleware(['auth', 'confirm_id'])->name('products.show');
